fix(model): cast softmax output to input dtype for FSDP mixed precision
F.softmax upscasts to float32 for numerical stability, but under FSDP
mixed precision the values tensor stays in bfloat16. This causes a
dtype mismatch at torch.matmul(attn, v). Fix by casting attn back to
v.dtype after softmax in both GQAttention and MLAttention.

Also revert bsub to always use torchrun and set micro_batch=4.

+++ open_mythos/main.py
@@ -242,7 +242,7 @@ class GQAttention(nn.Module):
         attn = torch.matmul(q, k.transpose(-2, -1)) * scale
         if mask is not None:
             attn = attn + mask
-        attn = self.attn_drop(F.softmax(attn, dim=-1))
+        attn = self.attn_drop(F.softmax(attn, dim=-1).to(v.dtype))
         out = torch.matmul(attn, v)
         out = out.transpose(1, 2).contiguous().view(B, T, -1)
         return self.wo(out)
@@ -384,7 +384,7 @@ class MLAttention(nn.Module):
         attn = torch.matmul(q, k.transpose(-2, -1)) * scale
         if mask is not None:
             attn = attn + mask
-        attn = self.attn_drop(F.softmax(attn, dim=-1))
+        attn = self.attn_drop(F.softmax(attn, dim=-1).to(v.dtype))
         out = torch.matmul(attn, v)  # (B, H, T, v_dim)
         out = out.transpose(1, 2).contiguous().view(B, T, -1)
         return self.wo(out)
